requirements.txt
guildai==0.6.6
numpy==1.18.1
scipy==1.4.1
tensorflow-gpu==2.0.1
tqdm==4.43.0
